Add files via upload
A simple script for working with Excel files (.xlsx). It has 3 options:
(1) File Data: for a specified file, it shows the number of sheets, their names and their number of cells in use (as a measure of "sheet size")
(2) Create Files: in a specified directory, it creates files with the names given by the user
(3) Create Sheets: in a specified directory, it creates a new file with the sheets' names given by the user

As it is still a simple script, it's mandatory to be careful with the format of the data given by the user.